Rename grader validator file name.
